+++ R/stat_dotsinterval.R
@@ -37,12 +37,11 @@ dots_sample_slab_function = function(
 # slab functions for distributions -------------------------
 
 #' @importFrom stats ppoints
-#' @importFrom purrr pmap_dfr
 dots_dist_slab_function = function(
   df, input, quantiles = 100, trans = scales::identity_trans(), ...
 ) {
-  pmap_dfr(df, function(dist, ...) {
-    if (is.na(dist)) {
+  pmap_dfr_(df, function(dist, ...) {
+    if (is.null(dist) || any(is.na(dist))) {
       return(data.frame(.input = NA, .value = NA))
     }
 
@@ -50,7 +49,7 @@ dots_dist_slab_function = function(
       list(p = ppoints(quantiles, a = 1/2)),
       args_from_aes(...)
     )
-    quantile_fun = match.fun(paste0("q", dist))
+    quantile_fun = dist_quantile_fun(dist)
     input = do.call(quantile_fun, args)
 
     data.frame(
@@ -210,7 +209,6 @@ stat_dist_dotsinterval = function(
     )
   )
 }
-#' @importFrom plyr defaults
 StatDistDotsinterval = ggproto("StatDistDotsinterval", StatDistSlabinterval,
   extra_params = c(
     StatDistSlabinterval$extra_params,